Create an invoice on the bitpay site through the Spree order
This is a refactoring of some code that was in the controller in v1,
but genuinely belongs in the model instead. Should improve the overall
design of the code.

This is the basic functionality to create an invoice and display it.

There are still possible problems with displaying invoices and finding
IPNS.